pyproject.toml
[project]
name = "procast-ai"
version = "0.1.0"
description = "AI-powered budget analysis agent for Procast event planning platform"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Procast Team"}
]
keywords = ["ai", "agent", "budget", "analysis", "langgraph", "dspy", "mcp"]

dependencies = [
    # Core LLM & Agent Framework
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langgraph>=0.2.0",
    "dspy-ai>=2.5.0",
    
    # MCP (Model Context Protocol)
    "mcp>=1.0.0",
    
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    
    # API Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    
    # SQL Parsing & Validation
    "sqlglot>=25.0.0",
    
    # Utilities
    "tenacity>=9.0.0",
    "structlog>=24.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    
    # Observability
    "langsmith>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "pre-commit>=3.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=term-missing"
